+++ plugin.video.iplayerwww/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.iplayerwww" name="iPlayer WWW" version="4.0.16" provider-name="CaptainT, Cas, ihurst, primaeval">
+<addon id="plugin.video.iplayerwww" name="iPlayer WWW" version="4.0.17" provider-name="CaptainT, Cas, ihurst, primaeval">
   <requires>
     <import addon="xbmc.python" version="3.0.0"/>
     <import addon="script.module.requests" version="2.7.0"/>
@@ -22,8 +22,8 @@
         <fanart>resources/fanart.jpg</fanart>
     </assets>
     <news>
-v4.0.16
-Live radio now needs an authentication token. Credits to dimkroon/kereltje once again.
+v4.0.17
+Live radio authentication token fixes and simplifications.
 
 Known issues:
 - No Red Button
