+++ plugin.video.iplayerwww/default.py
@@ -87,150 +87,153 @@ try:
 except:
     pass
 
-# These are the modes which tell the plugin where to go.
-if mode == 1:
-    Common.KidsMode()
+try:
+    # These are the modes which tell the plugin where to go.
+    if mode == 1:
+        Common.KidsMode()
 
-elif mode is None or url is None or len(url) < 1:
-    Common.CreateBaseDirectory(content_type)
+    elif mode is None or url is None or len(url) < 1:
+        Common.CreateBaseDirectory(content_type)
 
-# Modes 101-119 will create a main directory menu entry
-elif mode == 101:
-    Video.ListLive()
+    # Modes 101-119 will create a main directory menu entry
+    elif mode == 101:
+        Video.ListLive()
 
-elif mode == 102:
-    Video.ListAtoZ()
+    elif mode == 102:
+        Video.ListAtoZ()
 
-elif mode == 103:
-    Video.ListCategories()
+    elif mode == 103:
+        Video.ListCategories()
 
-elif mode == 104:
-    Video.Search(keyword)
+    elif mode == 104:
+        Video.Search(keyword)
 
-elif mode == 105:
-    Video.ListMostPopular()
+    elif mode == 105:
+        Video.ListMostPopular()
 
-elif mode == 106:
-    Video.ListHighlights(url)
+    elif mode == 106:
+        Video.ListHighlights(url)
 
-elif mode == 107:
-    Video.ListWatching()
+    elif mode == 107:
+        Video.ListWatching()
 
-elif mode == 108:
-    Video.ListFavourites()
+    elif mode == 108:
+        Video.ListFavourites()
 
-elif mode == 109:
-    Video.ListChannelHighlights()
+    elif mode == 109:
+        Video.ListChannelHighlights()
 
-elif mode == 112:
-    Radio.ListAtoZ()
+    elif mode == 112:
+        Radio.ListAtoZ()
 
-elif mode == 113:
-    Radio.ListLive()
+    elif mode == 113:
+        Radio.ListLive()
 
-elif mode == 114:
-    Radio.ListGenres()
+    elif mode == 114:
+        Radio.ListGenres()
 
-elif mode == 115:
-    Radio.Search(keyword)
+    elif mode == 115:
+        Radio.Search(keyword)
 
-elif mode == 116:
-    Radio.ListMostPopular()
+    elif mode == 116:
+        Radio.ListMostPopular()
 
-elif mode == 117:
-    Radio.ListListenList()
+    elif mode == 117:
+        Radio.ListListenList()
 
-elif mode == 199:
-    Radio.ListFollowing()
+    elif mode == 199:
+        Radio.ListFollowing()
 
-elif mode == 118:
-    Video.RedButtonDialog()
+    elif mode == 118:
+        Video.RedButtonDialog()
 
-elif mode == 119:
-    Common.SignOutBBCiD()
+    elif mode == 119:
+        Common.SignOutBBCiD()
 
-elif mode == 120:
-    Video.ListChannelAtoZ()
+    elif mode == 120:
+        Video.ListChannelAtoZ()
 
-    # Modes 121-199 will create a sub directory menu entry
-elif mode == 121:
-    Video.GetEpisodes(url)
+        # Modes 121-199 will create a sub directory menu entry
+    elif mode == 121:
+        Video.GetEpisodes(url)
 
-elif mode == 122:
-    Video.GetAvailableStreams(name, url, iconimage, description)
+    elif mode == 122:
+        Video.GetAvailableStreams(name, url, iconimage, description)
 
-elif mode == 123:
-    Video.AddAvailableLiveStreamsDirectory(name, url, iconimage)
+    elif mode == 123:
+        Video.AddAvailableLiveStreamsDirectory(name, url, iconimage)
 
-elif mode == 124:
-    Video.GetAtoZPage(url)
+    elif mode == 124:
+        Video.GetAtoZPage(url)
 
-elif mode == 125:
-    Video.ListCategoryFilters(url)
+    elif mode == 125:
+        Video.ListCategoryFilters(url)
 
-elif mode == 126:
-    Video.GetFilteredCategory(url)
+    elif mode == 126:
+        Video.GetFilteredCategory(url)
 
-elif mode == 127:
-    Video.GetGroup(url)
+    elif mode == 127:
+        Video.GetGroup(url)
 
-elif mode == 128:
-    Video.ScrapeEpisodes(url)
+    elif mode == 128:
+        Video.ScrapeEpisodes(url)
 
-elif mode == 129:
-    Video.AddAvailableRedButtonDirectory(name, url)
+    elif mode == 129:
+        Video.AddAvailableRedButtonDirectory(name, url)
 
-elif mode == 131:
-    Radio.GetEpisodes(url)
+    elif mode == 131:
+        Radio.GetEpisodes(url)
 
-elif mode == 132:
-    Radio.GetAvailableStreams(name, url, iconimage, description)
+    elif mode == 132:
+        Radio.GetAvailableStreams(name, url, iconimage, description)
 
-elif mode == 133:
-    Radio.AddAvailableLiveStreamsDirectory(name, url, iconimage)
+    elif mode == 133:
+        Radio.AddAvailableLiveStreamsDirectory(name, url, iconimage)
 
-elif mode == 134:
-    Video.ScrapeAtoZEpisodes(url)
+    elif mode == 134:
+        Video.ScrapeAtoZEpisodes(url)
 
-elif mode == 136:
-    Radio.GetPage(url)
+    elif mode == 136:
+        Radio.GetPage(url)
 
-elif mode == 137:
-    Radio.GetCategoryPage(url)
+    elif mode == 137:
+        Radio.GetCategoryPage(url)
 
-elif mode == 138:
-    Radio.GetAtoZPage(url)
+    elif mode == 138:
+        Radio.GetAtoZPage(url)
 
-elif mode == 139:
-    Video.ScrapeEpisodes(url)
+    elif mode == 139:
+        Video.ScrapeEpisodes(url)
 
-# Modes 201-299 will create a playable menu entry, not a directory
-elif mode == 201:
-    Video.PlayStream(name, url, iconimage, description, subtitles_url)
+    # Modes 201-299 will create a playable menu entry, not a directory
+    elif mode == 201:
+        Video.PlayStream(name, url, iconimage, description, subtitles_url)
 
-elif mode == 202:
-    Video.AddAvailableStreamItem(name, url, iconimage, description)
+    elif mode == 202:
+        Video.AddAvailableStreamItem(name, url, iconimage, description)
 
-elif mode == 203:
-    Video.AddAvailableLiveStreamItemSelector(name, url, iconimage)
+    elif mode == 203:
+        Video.AddAvailableLiveStreamItemSelector(name, url, iconimage)
 
-elif mode == 204:
-    Video.AddAvailableRedButtonItem(name, url)
+    elif mode == 204:
+        Video.AddAvailableRedButtonItem(name, url)
 
-elif mode == 205:
-    Video.AddAvailableUHDTrialItem(name, url)
+    elif mode == 205:
+        Video.AddAvailableUHDTrialItem(name, url)
 
-elif mode == 211:
-    Radio.PlayStream(name, url, iconimage, description, subtitles_url)
+    elif mode == 211:
+        Radio.PlayStream(name, url, iconimage, description, subtitles_url)
 
-elif mode == 212:
-    Radio.AddAvailableStreamItem(name, url, iconimage, description)
+    elif mode == 212:
+        Radio.AddAvailableStreamItem(name, url, iconimage, description)
 
-elif mode == 213:
-    Radio.AddAvailableLiveStreamItem(name, url, iconimage)
+    elif mode == 213:
+        Radio.AddAvailableLiveStreamItem(name, url, iconimage)
 
-elif mode == 197:
-    Video.ListUHDTrial()
-
+    elif mode == 197:
+        Video.ListUHDTrial()
+except Common.IpwwwError as err:
+    xbmcgui.Dialog().ok(Common.translation(30400), str(err))
+    sys.exit(1)
 
 xbmcplugin.endOfDirectory(int(sys.argv[1]))
